fix(palette): update files with config for palette to split the logic, simplify colors
