+++ src/components/Icon/Icon.stories.mdx
@@ -82,7 +82,7 @@ Icon with size.
   <Story name="Icon with size">
     <Stack>
     <PresentComponent name={'default'}>
-      <Icon name={'add'} />
+      <Icon name={'add'} color={'#de5798'} />
     </PresentComponent>
     <PresentComponent name={'20'}>
       <Icon name={'add'} size={20} />
