+++ src/components/Icon/Icon.style.ts
@@ -4,7 +4,7 @@ import { AcceptedColorComponentTypes, fillPickerBasedOnType } from '../../utils/
 
 type iconStyleProps = {
   /** Property indicating the color of the icon. Defaults to primary */
-  color: AcceptedColorComponentTypes;
+  color: AcceptedColorComponentTypes | string;
   /** Property indicating the size of the icon. Defaults to 16 */
   size: number;
 };
