+++ src/components/Icon/Icon.tsx
@@ -10,7 +10,7 @@ export type Props = {
   /** This property defines witch icon to use */
   name: AcceptedIconNames;
   /** Property indicating the color of the icon. Defaults to primary */
-  color?: AcceptedColorComponentTypes;
+  color?: AcceptedColorComponentTypes | string;
   /** Property indicating the size of the icon. Defaults to 16 */
   size?: number;
 };
