+++ src/utils/themeFunctions.ts
@@ -4,26 +4,13 @@ export type AcceptedColorComponentTypes =
   | 'primary'
   | 'secondary'
   | 'success'
-  | 'teal'
   | 'error'
   | 'info'
   | 'warning'
-  | 'light'
-  | 'dark'
   | 'branded1'
-  | 'branded2'
-  | 'lightGray100'
-  | 'lightGray200'
-  | 'lightGray400'
-  | 'lightGray700'
-  | 'darkGray400'
-  | 'darkGray600';
+  | 'branded2';
 
-const getColorFromType = (
-  type: AcceptedColorComponentTypes,
-  defaultValue: string,
-  theme: Theme
-) => {
+const getColorFromType = (type: AcceptedColorComponentTypes | string, theme: Theme) => {
   switch (type) {
     case 'primary':
       return theme.palette.primary[100];
@@ -31,8 +18,6 @@ const getColorFromType = (
       return theme.palette.secondary[400];
     case 'success':
       return theme.palette.success[400];
-    case 'teal':
-      return theme.palette.flat.teal[400];
     case 'error':
       return theme.palette.error[400];
     case 'info':
@@ -43,25 +28,13 @@ const getColorFromType = (
       return theme.palette.branded1[400];
     case 'branded2':
       return theme.palette.branded2[400];
-    case 'lightGray100':
-      return theme.palette.flat.lightGray[100];
-    case 'dark':
-      return theme.palette.flat.darkGray[600];
-    case 'light':
-      return theme.palette.text.light[400];
-    case 'lightGray400':
-      return theme.palette.flat.lightGray[400];
-    case 'lightGray700':
-      return theme.palette.flat.lightGray[700];
-    case 'darkGray400':
-      return theme.palette.flat.darkGray[400];
     default:
-      return defaultValue;
+      return type;
   }
 };
 
 export const backgroundPickerBasedOnType = (type: AcceptedColorComponentTypes) => (theme: Theme) =>
-  getColorFromType(type, theme.palette.flat.lightGray[100], theme);
+  getColorFromType(type, theme);
 
 export const colorPickerBasedOnType = (type: AcceptedColorComponentTypes) => (theme: Theme) => {
   switch (type) {
@@ -74,5 +47,6 @@ export const colorPickerBasedOnType = (type: AcceptedColorComponentTypes) => (th
   }
 };
 
-export const fillPickerBasedOnType = (type: AcceptedColorComponentTypes) => (theme: Theme) =>
-  getColorFromType(type, theme.palette.primary[100], theme);
+export const fillPickerBasedOnType = (type: AcceptedColorComponentTypes | string) => (
+  theme: Theme
+) => getColorFromType(type, theme);
